+++ tests/cli/agent/test_agent_attack.py
@@ -1,5 +1,6 @@
 #  Copyright 2021, Verizon Media
 #  Licensed under the terms of the ${MY_OSI} license. See the LICENSE file in the project root for terms
+import tempfile
 from argparse import Namespace
 from pathlib import Path
 from unittest import TestCase
@@ -18,6 +19,9 @@ class TestAgentAttackCLI(TestCase):
             Path(__file__).joinpath("../../../resources/testplans").resolve()
         )
         Coordinator.DEFAULT_DURATION = 1
+        self.attack_report_yaml_path: Path = Path(
+            tempfile.NamedTemporaryFile("w+", delete=False).name
+        )
 
     def test_valid_test_plan(self):
         args = Namespace()
@@ -25,6 +29,22 @@ class TestAgentAttackCLI(TestCase):
 
         # Required Arguments for TestPlanValidatorCommand
         args.testplan = self.test_plans_directory.joinpath("valid/testplan1.json")
+        args.attack_report_yaml = self.attack_report_yaml_path
+
+        # Create a Mocked CLI App
+        app = MockApp(args)
+        args.app = app
+        args.cls.main(args)
+        self.assertEqual(0, args.cls.main(args))
+        self.assertTrue(Path.is_file(self.attack_report_yaml_path))
+
+    def test_no_attack_report_path(self):
+        args = Namespace()
+        args.cls = self.cls
+
+        # Required Arguments for TestPlanValidatorCommand
+        args.testplan = self.test_plans_directory.joinpath("valid/testplan1.json")
+        args.attack_report_yaml = None
 
         # Create a Mocked CLI App
         app = MockApp(args)
@@ -32,12 +52,27 @@ class TestAgentAttackCLI(TestCase):
 
         self.assertEqual(0, args.cls.main(args))
 
+    def test_invalid_attack_report_path(self):
+        args = Namespace()
+        args.cls = self.cls
+
+        # Required Arguments for TestPlanValidatorCommand
+        args.testplan = self.test_plans_directory.joinpath("valid/testplan1.json")
+        args.attack_report_yaml = Path(".")
+
+        # Create a Mocked CLI App
+        app = MockApp(args)
+        args.app = app
+
+        self.assertEqual(1, args.cls.main(args))
+
     def test_invalid_test_plan(self):
         args = Namespace()
         args.cls = self.cls
 
         # Required Arguments for TestPlanValidatorCommand
         args.testplan = self.test_plans_directory.joinpath("invalid/testplan1.yaml")
+        args.attack_report_yaml = self.attack_report_yaml_path
 
         # Create a Mocked CLI App
         app = MockApp(args)
@@ -51,6 +86,7 @@ class TestAgentAttackCLI(TestCase):
 
         # Required Arguments for TestPlanValidatorCommand
         args.testplan = self.test_plans_directory.joinpath("invalid/invalid_path.yaml")
+        args.attack_report_yaml = self.attack_report_yaml_path
 
         # Create a Mocked CLI App
         app = MockApp(args)
@@ -64,6 +100,7 @@ class TestAgentAttackCLI(TestCase):
 
         # Required Arguments for TestPlanValidatorCommand
         args.testplan = self.test_plans_directory.joinpath("valid/")
+        args.attack_report_yaml = self.attack_report_yaml_path
 
         # Create a Mocked CLI App
         app = MockApp(args)